requirements.txt
# Django and core dependencies
Django==5.0.7

# Web scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1

# Environment configuration
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-django==4.7.0